+++ src/4_components/PageInfo.vue
@@ -22,7 +22,7 @@ export default {
 
 <style lang="sass" scoped>
 
-  @import "src/assets/styles/style-variables.sass"
+  @import "src/0_assets/styles/style-variables.sass"
 
   .page-description
     margin-bottom: 1rem
