+++ src/4_components/navigation/CtaButton.vue
@@ -23,7 +23,7 @@ export default {
 </script>
 <style lang="sass" scoped>
 
-  @import "src/assets/styles/style-variables.sass"
+  @import "src/0_assets/styles/style-variables.sass"
 
   /* CTA BUTTON STYLING */
