+++ src/4_components/profiles/SocialmediaItem.vue
@@ -25,7 +25,7 @@ export default {
 
 <style lang="sass" scoped>
 
-  @import "src/assets/styles/style-variables.sass"
+  @import "src/0_assets/styles/style-variables.sass"
 
   .social-media-item
     display: inline-block
